Corrected comparisons performed on memory addresses (pointers)
As it turns out, the non-deterministic behavior of the code was due to sorting of containers of pointers (to geometric objects) based on the actual addresses in memory. As there is no guarantee whatsoever where objects are created on the heap, this is not a bright idea. In all places (that I could find) where this was being done, the code now uses the IDs associated with every geometric object. It seems to be deterministic now - in the sense that it will produce reproducible outcomes.