Fixed rayTriangle intersection function
When the mesh was getting big enough for partial flushing during rendering, EXC_BAD_ACCESS exceptions frequently occurred. This was due to the usage of the default comparison for pointers used in std::set_intersection() and std::set_difference() instead of the ID-based comparison used to sort the sets being processed.