Two improvements.
(1) Intersection output. Instead of only displaying internal IDs of intersection sections (which can not be identified in the SWC file due to preconditioning), the output will now also include coordinates (which are a tad more useful; however, they still cannot be identified in the SWC file due to re-centering to soma system -- why is that necessary anyway?).
(2) Introduced a command line paramater for the refinement level of the soma sphere. This needs to be a user-defined paramter at the moment as it is not determined automatically depending on the sizes of the root neurite segments (which would be ideal as the smallest of them seems to define how well-refined the soma should be, otherwise, bad connections are created during post-processing).